--- Example/QuickLayout/Images.xcassets/close_icon.imageset/Contents.json
@@ -1,23 +1,0 @@
-{
-  "images" : [
-    {
-      "idiom" : "universal",
-      "filename" : "ic_close_36pt.png",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "ic_close_36pt_2x.png",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "filename" : "ic_close_36pt_3x.png",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
