+++ Example/QuickLayout/NavigationController.swift
@@ -3,7 +3,7 @@
 //  DemoApp
 //
 //  Created by Daniel Huri on 5/11/18.
-//  Copyright © 2018 CocoaPods. All rights reserved.
+//  Copyright (c) 2018 huri000@gmail.com. All rights reserved.
 //
 
 import UIKit
@@ -12,6 +12,7 @@ class NavigationController: UINavigationController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
+        navigationBar.tintColor = QLColor.BlueGray.c700
         pushViewController(MainViewController(), animated: false)
     }
 }
