+++ Example/QuickLayout/Samples/ScrollView/ScrollSampleViewController.swift
@@ -3,7 +3,7 @@
 //  QuickLayout_Example
 //
 //  Created by Daniel Huri on 11/21/17.
-//  Copyright © 2017 CocoaPods. All rights reserved.
+//  Copyright (c) 2018 huri000@gmail.com. All rights reserved.
 //
 
 import UIKit
