+++ Example/QuickLayout/Samples/TableView/Contact.swift
@@ -1,0 +1,17 @@
+//
+//  Contact.swift
+//  DemoApp
+//
+//  Created by Daniel Huri on 5/11/18.
+//  Copyright (c) 2018 huri000@gmail.com. All rights reserved.
+//
+
+import Foundation
+
+
+struct Contact: Codable {
+    var firstName: String
+    var lastName: String
+    var userName: String
+    var email: String
+}
