+++ README.md
@@ -102,7 +102,7 @@ Swift 4.0 and iOS 9.0 (or higher).
     
 #### CocoaPods
 ```
-pod 'QuickLayout', '1.0.13'
+pod 'QuickLayout', '1.0.14'
 ```
 
 #### Manually
